split out the handlers
